Change coeff to percentage
Modified flock parameters type from integer to float

+++ Assets/Scripts/BirdBehavior/ClingyBehavior.cs
@@ -1,14 +1,13 @@
 using UnityEngine;
 
 // Collant : percute souvent ses camarades
-// Forte cohésion, faible séparation
 public class ClingyBehavior : BaseBoidBehavior
 {
-    public ClingyBehavior(FlockManager flockManager, int dense, int loose, int elongated) : base(flockManager, dense, loose, elongated) { }
+    public ClingyBehavior(FlockManager flockManager, float dense, float loose, float elongated) : 
+        base(flockManager, dense, loose, elongated) { }
 
     public override Color GetColor()
     {
         return Color.green;
     }
 }
-
