+++ Assets/Scripts/BirdBehavior/LatecomerBehavior.cs
@@ -1,10 +1,10 @@
 using UnityEngine;
 
 // Laggard: often isolated and falling behind
-// Low cohesion, high separation
 public class LatecomerBehavior : BaseBoidBehavior
 {
-    public LatecomerBehavior(FlockManager flockManager, int dense, int loose, int elongated) : base(flockManager, dense, loose, elongated) { }
+    public LatecomerBehavior(FlockManager flockManager, float dense, float loose, float elongated) : 
+        base(flockManager, dense, loose, elongated) { }
 
     public override Color GetColor()
     {
